requirements.backend.txt
llama-index
chromadb
llama-index-vector-stores-chroma
llama-index-embeddings-huggingface
llama-index-retrievers-bm25
python-dotenv
pypdf
uvicorn
fastapi
groq
requests
rank-bm25
sentence-transformers
huggingface_hub>=0.20.0
